Use graphql for concept search
