+++ src/containers/ListingPage/ListingPage.jsx
@@ -5,158 +5,24 @@
  * LICENSE file in the root directory of this source tree.
  *
  */
-import React, { useState, useEffect, Fragment } from 'react';
+import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
-import Helmet from 'react-helmet';
-import { Remarkable } from 'remarkable';
-import Downshift from 'downshift';
-import { css } from '@emotion/core';
-import styled from '@emotion/styled';
-import { colors, fonts, spacing } from '@ndla/core';
 import { injectT } from '@ndla/i18n';
-import ListView from '@ndla/listview';
-import {
-  OneColumn,
-  FilterListPhone,
-  LanguageSelector,
-  MastheadItem,
-  CreatedBy,
-} from '@ndla/ui';
-import { DropdownInput, DropdownMenu } from '@ndla/forms';
-import { ChevronDown, Search } from '@ndla/icons/lib/common';
-import Button from '@ndla/button';
-import { Spinner } from '@ndla/ui';
 
-import {
-  mapTagsToFilters,
-  mapConceptToListItem,
-} from '../../util/listingHelpers';
-import useQueryParameter from '../../util/useQueryParameter';
-import {
-  fetchConcept,
-  fetchConceptsBySubject,
-  fetchTags,
-  fetchConceptsByTags,
-  fetchConcepts,
-} from '../../api/concept/conceptApi';
+import { mapTagsToFilters } from '../../util/listingHelpers';
+import { fetchTags } from '../../api/concept/conceptApi';
 import { fetchSubjectIds, fetchSubject } from '../Subject/subjectApi';
-import { getLocaleUrls } from '../../util/localeHelpers';
-import ConceptPage from '../../components/Concept/ConceptPage';
-import Footer from '../../components/Footer';
-import CopyTextButton from '../../components/CopyTextButton';
-import config from '../../config';
-
-const SubjectFilterWrapper = styled.div`
-  margin-top: ${spacing.large};
-  margin-bottom: ${spacing.small};
-`;
-
-const HeaderWithLanguageWrapper = styled.div`
-  display: flex;
-`;
-
-const StyledLanguageSelector = styled.div`
-  margin-top: ${spacing.large};
-`;
-
-const StyledEmbedCopyButton = styled.div`
-  margin-top: ${spacing.large};
-  margin-left: auto;
-`;
-
-const SeparatorWrapper = styled.div`
-  margin-bottom: ${spacing.small};
-  padding-left: ${spacing.small};
-`;
-
-const CategoriesFilterWrapper = styled.div`
-  margin-bottom: ${spacing.small};
-  position: relative;
-  display: inline-block;
-`;
-
-const ButtonWrapper = styled.div`
-  display: flex;
-  justify-content: center;
-  margin: ${spacing.medium};
-`;
-
-const placeholderCSS = css`
-  color: initial;
-  font-weight: initial;
-  opacity: 0.5;
-`;
-const placeholderHasValuesCSS = props =>
-  !props.hasValues
-    ? css`
-        color: ${colors.brand.primary};
-        font-weight: bold;
-        ${fonts.sizes('16px')};
-      `
-    : placeholderCSS;
-
-const categoryFilterCSS = props => css`
-  border: 2px solid ${colors.brand.primary};
-  min-height: auto;
-  cursor: pointer;
-  background-color: transparent;
-  flex-grow: 0;
-  input {
-    cursor: pointer;
-    ::placeholder {
-      ${placeholderHasValuesCSS(props)}
-    }
-    :focus {
-      ::placeholder {
-        ${placeholderCSS}
-      }
-    }
-  }
-`;
-
-const PAGE_SIZE = 100;
+import ListingContainer from './ListingContainer';
 
 const ListingPage = ({ t, locale, location, isOembed }) => {
-  const [concepts, setConcepts] = useState([]);
-  const [totalCount, setTotalCount] = useState(0);
   const [subjectIds, setSubjectIds] = useState([]);
   const [subjects, setSubjects] = useState([]);
   const [filters, setFilters] = useState(new Map());
   const [tags, setTags] = useState([]);
-  const [page, setPage] = useState(1);
-  const [showButton, setShowButton] = useState(true);
-  const [loading, setLoading] = useState(false);
-  const [currentListFilters, setCurrentListFilters] = useState([]);
-  const [selectedListFilter, setSelectedListFilter] = useState(null);
-  const [viewStyle, setViewStyle] = useState('grid');
-  const [detailedItem, setDetailedItem] = useState(null);
-  const [selectedConcept, setSelectedConcept] = useState(null);
-  const [searchValue, setSearchValue] = useState('');
-  const [filterSearchValue, setFilterSearchValue] = useState('');
-  const [filterListOpen, setFilterListOpen] = useState(false);
-  const [queryParams, setQueryParams] = useQueryParameter({
-    subjects: [],
-    filters: [],
-    concept: null,
-  });
-  const [md, setMd] = useState(null);
-
-  useEffect(() => {
-    if (md === null) {
-      const markdown = new Remarkable();
-      markdown.inline.ruler.enable(['sub', 'sup']);
-      setMd(markdown);
-    }
-  }, []);
 
   useEffect(() => {
     getInitialData();
-    setSelectedListFilter(queryParams.filters?.[0]);
   }, []);
-
-  useEffect(() => {
-    getConceptFromQuery();
-  }, [queryParams.concept]);
 
   const getInitialData = async () => {
     const subjectIds = await fetchSubjectIds();
@@ -168,18 +34,10 @@ const ListingPage = ({ t, locale, location, isOembed }) => {
     setFilters(mapTagsToFilters(tags));
   };
 
-  const useDebounce = (val, delay) => {
-    const [debouncedVal, setDebouncedVal] = useState(val);
-    useEffect(() => {
-      const handler = setTimeout(() => {
-        setDebouncedVal(val);
-      }, delay);
-      return () => clearTimeout(handler);
-    }, [val, delay]);
-    return `${debouncedVal}*`;
-  };
-
-  const debouncedSearchVal = useDebounce(searchValue, 200);
+  /*
+  useEffect(() => {
+    getConceptFromQuery();
+  }, [queryParams.concept]);
 
   useEffect(() => {
     if (tags.length && subjectIds.length) {
@@ -247,299 +105,16 @@ const ListingPage = ({ t, locale, location, isOembed }) => {
     }
     setTotalCount(totalCount);
   };
-
-  const handleChangeSubject = values => {
-    setQueryParams({
-      subjects: values,
-      filters: [],
-    });
-    setSelectedListFilter(null);
-    setPage(1);
-  };
-
-  const handleChangeFilters = (key, values) => {
-    setQueryParams({
-      ...queryParams,
-      filters: values,
-    });
-    setPage(1);
-  };
-
-  const handleSelectItem = value => {
-    setSelectedConcept(value);
-    setQueryParams({
-      ...queryParams,
-      concept: value?.id,
-    });
-  };
-
-  const onConceptSearch = async value => {
-    setSearchValue(value);
-    setShowButton(!value);
-  };
-
-  const handleChangeListFilter = value => {
-    setQueryParams({
-      subjects: [],
-      filters: [value],
-    });
-    setFilterListOpen(false);
-    setSelectedListFilter(value);
-    setPage(1);
-  };
-
-  const handleStateChangeListFilter = changes => {
-    const { isOpen, type } = changes;
-
-    if (type === Downshift.stateChangeTypes.mouseUp) {
-      setFilterListOpen(isOpen);
-      if (!isOpen) {
-        setFilterSearchValue('');
-      }
-    }
-
-    if (type === Downshift.stateChangeTypes.keyDownEnter) {
-      setFilterSearchValue('');
-    }
-  };
-
-  const handleRemoveFilter = value => {
-    setQueryParams({
-      ...queryParams,
-      filters: [],
-    });
-    setSelectedListFilter(null);
-  };
-
-  const onFilterSearch = e => {
-    const {
-      target: { value },
-    } = e;
-    const filteredFilters = Array.from(filters.keys()).filter(item =>
-      item.toLowerCase().startsWith(value.toLowerCase()),
-    );
-    setFilterSearchValue(value);
-    setCurrentListFilters(filteredFilters);
-  };
-
-  const onFilterSearchFocus = () => {
-    setFilterListOpen(true);
-    setCurrentListFilters(Array.from(filters.keys()));
-  };
-
-  const onLoadMoreClick = () => {
-    getConcepts(page + 1);
-    setPage(page + 1);
-  };
-
-  const filterItems = listItems => {
-    let filteredItems = listItems;
-    if (queryParams.filters.length) {
-      filteredItems = filteredItems.filter(item =>
-        queryParams.filters.every(filter => item.filters.includes(filter)),
-      );
-    }
-    return filteredItems;
-  };
-
-  const getFilters = () => {
-    return filters.get(selectedListFilter)
-      ? [
-          {
-            filterValues: queryParams.filters,
-            onChange: handleChangeFilters,
-            isGroupedOptions: true,
-            key: 'default',
-            label: 'Filter',
-            options: [
-              filters.get(selectedListFilter).main.map(filter => ({
-                title: filter,
-                value: filter,
-                disabled: !listItems.some(item =>
-                  item.filters.includes(filter),
-                ),
-              })),
-              filters.get(selectedListFilter).sub.map(filter => ({
-                title: filter,
-                value: filter,
-                disabled: !listItems.some(item =>
-                  item.filters.includes(filter),
-                ),
-              })),
-            ],
-          },
-        ]
-      : [];
-  };
-
-  const getEmbedCode = (domain, filter) => {
-    return `<iframe aria-label="${filter}" src="${domain}/listing?filters[]=${filter}" frameborder="0" allowFullscreen="" />`;
-  };
-
-  const renderMarkdown = text => {
-    const rendered = md.render(text);
-    return (
-      <Fragment>
-        <span dangerouslySetInnerHTML={{ __html: rendered }} />
-      </Fragment>
-    );
-  };
-
-  const listItems = filterItems(
-    concepts.map(concept => mapConceptToListItem(concept)),
-  );
-
-  const categoryFilterInputProps = {
-    value: filterSearchValue,
-    onChange: onFilterSearch,
-    onFocus: onFilterSearchFocus,
-    onClick: onFilterSearchFocus,
-    placeholder: t(`listview.filters.category.openFilter`),
-  };
+  */
 
   return (
-    <>
-      <OneColumn>
-        <Helmet title={t(`themePage.heading`)} />
-        {!isOembed && (
-          <>
-            <HeaderWithLanguageWrapper>
-              <SubjectFilterWrapper>
-                <FilterListPhone
-                  preid="subject-list"
-                  label={t(`listview.filters.subject.openFilter`)}
-                  options={subjects.map(item => ({
-                    title: item.name,
-                    value: item.id,
-                  }))}
-                  alignedGroup
-                  showActiveFiltersOnSmallScreen
-                  values={queryParams.subjects}
-                  messages={{
-                    useFilter: t(`listview.filters.subject.useFilter`),
-                    openFilter: t(`listview.filters.subject.openFilter`),
-                    closeFilter: t(`listview.filters.subject.closeFilter`),
-                  }}
-                  onChange={handleChangeSubject}
-                  viewMode="allModal"
-                />
-              </SubjectFilterWrapper>
-              <StyledEmbedCopyButton>
-                {selectedListFilter && (
-                  <CopyTextButton
-                    copyTitle={t('listview.embedlink.copyTitle')}
-                    hasCopiedTitle={t('listview.embedlink.hasCopiedTitle')}
-                    stringToCopy={getEmbedCode(
-                      config.ndlaListingFrontendDomain,
-                      selectedListFilter,
-                    )}
-                    timeout={5000}
-                    ghostPill
-                  />
-                )}
-              </StyledEmbedCopyButton>
-              <StyledLanguageSelector>
-                <MastheadItem>
-                  <LanguageSelector
-                    options={getLocaleUrls(locale, location)}
-                    currentLanguage={locale}
-                    alwaysVisible
-                  />
-                </MastheadItem>
-              </StyledLanguageSelector>
-            </HeaderWithLanguageWrapper>
-            <SeparatorWrapper>{t(`listingPage.or`)}</SeparatorWrapper>
-            <CategoriesFilterWrapper>
-              <Downshift
-                onSelect={handleChangeListFilter}
-                onStateChange={handleStateChangeListFilter}
-                isOpen={filterListOpen}>
-                {({ getInputProps, getMenuProps, getItemProps }) => {
-                  return (
-                    <div>
-                      <DropdownInput
-                        multiSelect
-                        {...getInputProps(categoryFilterInputProps)}
-                        data-testid={'dropdownInput'}
-                        iconRight={
-                          filterListOpen ? (
-                            <Search />
-                          ) : (
-                            <span onClick={onFilterSearchFocus}>
-                              <ChevronDown />
-                            </span>
-                          )
-                        }
-                        values={selectedListFilter ? [selectedListFilter] : []}
-                        removeItem={handleRemoveFilter}
-                        customCSS={categoryFilterCSS({
-                          hasValues: selectedListFilter,
-                        })}
-                      />
-                      <DropdownMenu
-                        getMenuProps={getMenuProps}
-                        getItemProps={getItemProps}
-                        isOpen={filterListOpen}
-                        items={currentListFilters}
-                        maxRender={1000}
-                        hideTotalSearchCount
-                        positionAbsolute
-                      />
-                    </div>
-                  );
-                }}
-              </Downshift>
-            </CategoriesFilterWrapper>
-          </>
-        )}
-        <ListView
-          disableSearch={isOembed}
-          disableViewOption={isOembed}
-          items={listItems}
-          detailedItem={detailedItem}
-          selectCallback={setDetailedItem}
-          viewStyle={viewStyle}
-          onChangedViewStyle={e => setViewStyle(e.viewStyle)}
-          searchValue={searchValue}
-          onChangedSearchValue={e => onConceptSearch(e.target.value)}
-          selectedItem={
-            selectedConcept ? (
-              <ConceptPage
-                conceptId={Number(selectedConcept.id)}
-                language={locale}
-                inModal={true}
-                handleClose={handleSelectItem}
-              />
-            ) : null
-          }
-          onSelectItem={handleSelectItem}
-          renderMarkdown={renderMarkdown}
-          filters={isOembed ? [] : getFilters()}
-          totalCount={totalCount}
-        />
-        {showButton && (
-          <ButtonWrapper>
-            {loading ? (
-              <Spinner />
-            ) : (
-              <Button onClick={onLoadMoreClick}>
-                {t('listingPage.loadMore')}
-              </Button>
-            )}
-          </ButtonWrapper>
-        )}
-        {isOembed && (
-          <CreatedBy
-            name={t('createdBy.listing.content')}
-            description={t('createdBy.listing.text')}
-            url={`${config.ndlaListingFrontendDomain}/${decodeURIComponent(
-              location.search,
-            )}`}
-          />
-        )}
-      </OneColumn>
-      {!isOembed && <Footer t={t} />}
-    </>
+    <ListingContainer
+      isOembed={isOembed}
+      subjects={subjects}
+      filters={filters}
+      location={location}
+      locale={locale}
+    />
   );
 };
